refactor(coordinator): extract ingestion manager
