+++ ingestion/__init__.py
@@ -1,0 +1,1 @@
+from .ingestion_manager import IngestionManager as IngestionManager
